+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### 0.4.0
+- Add the ability to specify multiple aliases for a permission.
+- Rename "message_if_false" -> "message_if_true" in PermissionResult.
+- Move "get_result_value" from manager to utils.
+
 ### 0.3.0
 - Get a parent_permission_manager from the `context` in the `PermissionManager` if it's passed.
 - Add `message_if_false` that indicating whether to include the `message` when the `value` is False.